requirements.txt
asgiref==3.4.1
Django==3.2.6
djangorestframework==3.12.4
psycopg2-binary==2.9.1
pytz==2021.1
sqlparse==0.4.1
django-rest-auth==0.9.5
django-allauth==0.45.0
factory-boy==3.2.0
python-dateutil==2.8.2
Faker==8.13.2
